refactor code placements
