+++ src/main/java/dk/kb/present/util/saxhandlers/FormHandler.java
@@ -1,10 +1,10 @@
-package dk.kb.present.saxhandlers;
+package dk.kb.present.util.saxhandlers;
 
 import org.xml.sax.Attributes;
 import org.xml.sax.SAXException;
 import org.xml.sax.helpers.DefaultHandler;
 
-public class ProductionCountryHandler extends DefaultHandler {
+public class FormHandler extends DefaultHandler {
     private StringBuilder currentValue = new StringBuilder();
     private String currentPath = "";
     private boolean captureValue = false;
@@ -15,7 +15,7 @@ public class ProductionCountryHandler extends DefaultHandler {
         currentPath += "/" + qName;
 
         // Check if the current path matches the target path
-        if ("/XIP/Metadata/Content/ns2:record/source/tvmeter/productioncountry".equals(currentPath)) {
+        if ("/XIP/Metadata/Content/ns2:record/source/tvmeter/form".equals(currentPath)) {
             captureValue = true;
         }
     }
@@ -23,7 +23,7 @@ public class ProductionCountryHandler extends DefaultHandler {
     @Override
     public void endElement(String uri, String localName, String qName) throws SAXException {
         // Check if we are at the end of the target element
-        if ("/XIP/Metadata/Content/ns2:record/source/tvmeter/productioncountry".equals(currentPath)) {
+        if ("/XIP/Metadata/Content/ns2:record/source/tvmeter/form".equals(currentPath)) {
             captureValue = false;
         }
